requirements.txt
python-telegram-bot==20.6
transformers==4.40.0
torch
Pillow
apscheduler
